笔记接口初步

+++ src/main/java/com/bulldog/monkey/api/model/UserEntity.java
@@ -1,6 +1,5 @@
 package com.bulldog.monkey.api.model;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -12,22 +11,22 @@ import io.swagger.annotations.ApiModelProperty;
 @ApiModel(description = "UserEntity 实体类")
 public class UserEntity {
     @ApiModelProperty(value = "用户id",dataType="int")
-    private int id;
+    private Long id;
     @ApiModelProperty(value = "用户名", required = true)
     private String name;
 
-    public UserEntity(@JsonProperty("id") int id,
+    public UserEntity(@JsonProperty("id") Long id,
                  @JsonProperty("name") String name) {
 
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
